requirements.txt
alembic==1.14.1
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
mysqlclient==2.2.7
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
python-dotenv==1.0.1
setuptools==75.8.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.2.1
